\name{getNetworkHost}
\alias{getNetworkHost}
\title{Returns the current network host configured for OLAPRGateway}
\usage{getNetworkHost()}
\arguments{
 
}
\value{
The resulting host as vector
}
\description{
Returns the current network host configured for OLAPRGateway
NOTE: OLAPRGateway.exe must be running to use this method. OLAPRGateway can be obtained here: https://github.com/jdeats76/OLAPRGateway 
}
\examples{
getNetworkPort()
}

