\name{getNetworkPort}
\alias{getNetworkPort}
\title{Returns the current network port configured for OLAPRGateway}
\usage{getNetworkPort()}
\arguments{
 
}
\value{
The resulting port number as vector
}
\description{
Returns the current network port configured for OLAPRGateway
NOTE: OLAPRGateway.exe must be running to use this method. OLAPRGateway can be obtained here: https://github.com/jdeats76/OLAPRGateway 
}
\examples{
getNetworkPort()
}

