\name{olaprgateway-package}
\alias{olaprgateway-package}
\alias{olaprgateway}
\docType{package}
\title{
\packageTitle{olaprgateway}
}
\description{
\packageDescription{olaprgateway}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{olaprgateway}
\packageIndices{olaprgateway}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{olaprgateway}

Maintainer: \packageMaintainer{olaprgateway}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
