\name{setOLAPGatewayPort}
\alias{setOLAPGatewayPort}
\title{sets the OLAPRGatewayPort to be used by the other methods in the OLAPRGateway package}
\usage{setOLAPGatewayPort(port)}
\arguments{
  \item{port}{The port to be used by OLAPRGateway methods. Default is 8888}
}
\value{
This method returns nothing
}
\description{
sets the OLAPRGatewayPort to be used by the other methods in the OLAPRGateway package
NOTE: OLAPRGateway.exe must be running to use this method. OLAPRGateway can be obtained here: https://github.com/jdeats76/OLAPRGateway 
}
\examples{
setOLAPRGatewayPort(8888)
}
 
