\name{validateTokenSet}
\alias{validateTokenSet}
\title{Confirms if the OLAPRGateway token has been set}
\usage{validateTokenSet()}
\arguments{
  
}
\value{
This method returns "True" or "False"
}
\description{
Validates if the OLAPRGateway token value has been set
NOTE: OLAPRGateway.exe must be running to use this method. OLAPRGateway can be obtained here: https://github.com/jdeats76/OLAPRGateway 
}
validateTOkenSet()
}
